Simplify sidebar
