+++ resources/js/Layouts/MainLayout.tsx
@@ -1,5 +1,5 @@
-import { AppSidebar } from "@/components/app-sidebar"
 import { MainBreadcrumb } from "@/components/nav/Breadcrumb"
+import { MainSidebar } from "@/components/nav/MainSidebar"
 import { ThemeProvider } from "@/components/ThemeProvider"
 import { Separator } from "@/components/ui/separator"
 import {
@@ -11,7 +11,7 @@ export default function MainLayout({ children }: { children: React.ReactNode }) 
     <ThemeProvider defaultTheme="dark" storageKey="vite-ui-theme">
       <div className="font-mono antialiased bg-background text-foreground">
         <SidebarProvider>
-          <AppSidebar />
+          <MainSidebar />
           <SidebarInset>
             <div className="fixed">
               <header className="flex h-14 shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-14">
