--- resources/js/components/app-sidebar.tsx
@@ -1,209 +1,0 @@
-import {
-  AudioWaveform, BookOpen, BookTextIcon, Bot, ChevronDown, Command, Frame,
-  GalleryVerticalEnd, HammerIcon, Map, PieChart, Settings2, SquareTerminal,
-  WrenchIcon
-} from "lucide-react"
-import * as React from "react"
-import IconOpenAgents from "@/components/IconOpenAgents"
-import { ModeToggle } from "@/components/ModeToggle"
-import { atlantisPortsChats, NavChats } from "@/components/nav-chats"
-import { NavMain } from "@/components/nav-main"
-import { NavProjects } from "@/components/nav-projects"
-import { NavUser } from "@/components/nav-user"
-import { TeamSwitcher } from "@/components/team-switcher"
-import {
-  Collapsible, CollapsibleContent, CollapsibleTrigger
-} from "@/components/ui/collapsible"
-import {
-  Sidebar, SidebarContent, SidebarFooter, SidebarGroup, SidebarGroupLabel,
-  SidebarHeader, SidebarRail, SidebarTrigger
-} from "@/components/ui/sidebar"
-
-// This is sample data.
-const data = {
-  user: {
-    name: "Christopher David",
-    email: "chris@openagents.com",
-    avatar: "https://pbs.twimg.com/profile_images/1607882836740120576/3Tg1mTYJ_400x400.jpg",
-  },
-  teams: [
-    {
-      name: "Atlantis Ports",
-      logo: GalleryVerticalEnd,
-      plan: "Pro",
-    },
-    {
-      name: "OpenAgents",
-      logo: AudioWaveform,
-      plan: "Startup",
-    },
-    {
-      name: "RoA",
-      logo: Command,
-      plan: "Free",
-    },
-  ],
-  navMain: [
-    {
-      title: "Playground",
-      url: "#",
-      icon: SquareTerminal,
-      isActive: true,
-      items: [
-        {
-          title: "History",
-          url: "#",
-        },
-        {
-          title: "Starred",
-          url: "#",
-        },
-        {
-          title: "Settings",
-          url: "#",
-        },
-      ],
-    },
-    {
-      title: "Models",
-      url: "#",
-      icon: Bot,
-      items: [
-        {
-          title: "Genesis",
-          url: "#",
-        },
-        {
-          title: "Explorer",
-          url: "#",
-        },
-        {
-          title: "Quantum",
-          url: "#",
-        },
-      ],
-    },
-    {
-      title: "Documentation",
-      url: "#",
-      icon: BookOpen,
-      items: [
-        {
-          title: "Introduction",
-          url: "#",
-        },
-        {
-          title: "Get Started",
-          url: "#",
-        },
-        {
-          title: "Tutorials",
-          url: "#",
-        },
-        {
-          title: "Changelog",
-          url: "#",
-        },
-      ],
-    },
-    {
-      title: "Settings",
-      url: "#",
-      icon: Settings2,
-      items: [
-        {
-          title: "General",
-          url: "#",
-        },
-        {
-          title: "Team",
-          url: "#",
-        },
-        {
-          title: "Billing",
-          url: "#",
-        },
-        {
-          title: "Limits",
-          url: "#",
-        },
-      ],
-    },
-  ],
-  projects: [
-    {
-      name: "Competitor Research",
-      url: "/",
-      icon: BookTextIcon,
-    },
-    {
-      name: "Container Port Design",
-      url: "/",
-      icon: WrenchIcon,
-    },
-    {
-      name: "Spaceport Design",
-      url: "/",
-      icon: WrenchIcon,
-    },
-    {
-      name: "Mining Rig Design",
-      url: "/",
-      icon: WrenchIcon,
-    },
-  ],
-}
-
-export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
-  return (
-    <Sidebar collapsible="icon" {...props}>
-      <SidebarHeader>
-        <div className="flex flex-col gap-2 py-2">
-          <div className="px-2 mb-4 flex items-center justify-between">
-            <div className="w-7">
-              <SidebarTrigger className="-ml-1" />
-            </div>
-            {/* <IconOpenAgents className="h-5 w-5" /> */}
-            <ModeToggle />
-          </div>
-          <TeamSwitcher teams={data.teams} />
-        </div>
-      </SidebarHeader>
-      <SidebarContent>
-
-        <Collapsible defaultOpen className="group/collapsible">
-          <SidebarGroup>
-            <SidebarGroupLabel asChild>
-              <CollapsibleTrigger>
-                Projects
-                <ChevronDown className="ml-auto transition-transform duration-200 group-data-[state=open]/collapsible:rotate-180" />
-              </CollapsibleTrigger>
-            </SidebarGroupLabel>
-            <CollapsibleContent>
-              <NavProjects projects={data.projects} />
-            </CollapsibleContent>
-          </SidebarGroup>
-        </Collapsible>
-
-        <Collapsible defaultOpen className="group/collapsible">
-          <SidebarGroup>
-            <SidebarGroupLabel asChild>
-              <CollapsibleTrigger>
-                Chats
-                <ChevronDown className="ml-auto transition-transform duration-200 group-data-[state=open]/collapsible:rotate-180" />
-              </CollapsibleTrigger>
-            </SidebarGroupLabel>
-            <CollapsibleContent>
-              <NavChats chats={atlantisPortsChats} highlightedChat="Portunus Project" />
-            </CollapsibleContent>
-          </SidebarGroup>
-        </Collapsible>
-
-      </SidebarContent>
-      <SidebarFooter>
-        <NavUser user={data.user} />
-      </SidebarFooter>
-      <SidebarRail />
-    </Sidebar>
-  )
-}
